requirements.txt
openai>=1.0.0
google-genai>=1.41.0
python-docx>=1.1.0
schedule>=1.2.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
docx2pdf>=0.1.8
-r job_scrape/requirements.txt
-r job_fit_analysis/requirements.txt
